[ch310] Added check for empty type before deletion

+++ app/Http/Controllers/TypeController.php
@@ -111,6 +111,12 @@ class TypeController extends Controller
 					]);
 		$type = Auth::user()->types()->find($validated['id']);
 
+		if ($type->entries->count() > 0) {
+			return response()->json([
+				'error' => true, 'message' => 'Type not empty!'
+			]);
+		}
+
 		if ($type) {
 			$type->delete();
